LPS-128289 Remove testGetLinkToLayoutValue from JournalConverterUtilTest and replace it with testGetLinkToLayoutValue in JournalContentCompatibilityConverterTest
